Modify folder and package structure
