+++ apps/android/app/src/main/kotlin/com/taskodoro/android/app/di/DiContainer.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -18,10 +18,10 @@ package com.taskodoro.android.app.di
 
 import android.content.Context
 import com.taskodoro.onboarding.OnboardingStoreFactory
-import com.taskodoro.storage.tasks.LocalTaskRepository
-import com.taskodoro.storage.tasks.TaskStoreFactory
-import com.taskodoro.tasks.new.TaskNew
-import com.taskodoro.tasks.new.TaskNewUseCase
+import com.taskodoro.tasks.data.TaskRepository
+import com.taskodoro.tasks.data.TaskStoreFactory
+import com.taskodoro.tasks.feature.new.TaskNew
+import com.taskodoro.tasks.feature.new.TaskNewUseCase
 import com.taskodoro.tasks.validator.TaskValidatorFactory
 import java.time.Instant
 import java.time.ZoneId
@@ -44,6 +44,6 @@ class DiContainer(
         }
 
     private val taskRepository by lazy {
-        LocalTaskRepository(TaskStoreFactory(applicationContext).create())
+        TaskRepository(TaskStoreFactory(applicationContext).create())
     }
 }
