+++ apps/android/app/src/main/kotlin/com/taskodoro/android/app/navigation/graphs/tasks/TaskNavigationGraph.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package com.taskodoro.android.app.navigation.graphs.tasks
 
-import com.taskodoro.tasks.new.TaskNewUseCase
+import com.taskodoro.tasks.feature.new.TaskNewUseCase
 import moe.tlaster.precompose.navigation.NavOptions
 import moe.tlaster.precompose.navigation.Navigator
 import moe.tlaster.precompose.navigation.RouteBuilder
