+++ apps/android/app/src/main/kotlin/com/taskodoro/android/app/navigation/graphs/tasks/TaskNewNavigation.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import com.taskodoro.android.app.tasks.new.TaskNewScreen
 import com.taskodoro.android.app.tasks.new.TaskNewViewModel
-import com.taskodoro.tasks.new.TaskNewUseCase
+import com.taskodoro.tasks.feature.new.TaskNewUseCase
 import kotlinx.coroutines.Dispatchers
 import moe.tlaster.precompose.navigation.BackHandler
 import moe.tlaster.precompose.navigation.NavOptions
