+++ taskodoro/src/commonMain/kotlin/com/taskodoro/onboarding/OnboardingStoreFactory.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
  */
 
 package com.taskodoro.onboarding
+
+import com.taskodoro.onboarding.feature.OnboardingStore
 
 expect class OnboardingStoreFactory {
     fun create(): OnboardingStore
