+++ taskodoro/src/commonMain/kotlin/com/taskodoro/tasks/data/TaskStoreFactory.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,21 +14,8 @@
  *    limitations under the License.
  */
 
-package com.taskodoro.storage.tasks
+package com.taskodoro.tasks.data
 
-import android.content.Context
-import com.taskodoro.storage.db.DatabaseFactory
-import com.taskodoro.storage.tasks.store.SQLDelightTaskStore
-
-actual class TaskStoreFactory(
-    private val context: Context,
-) {
-
-    actual fun create(): TaskStore {
-        val database = DatabaseFactory(context).create().apply {
-            taskdoroDBQueries.clearDB()
-        }
-
-        return SQLDelightTaskStore(database = database)
-    }
+expect class TaskStoreFactory {
+    fun create(): TaskStore
 }
