+++ taskodoro/src/commonMain/kotlin/com/taskodoro/tasks/data/local/SQLDelightTaskStore.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  *    limitations under the License.
  */
 
-package com.taskodoro.storage.tasks.store
+package com.taskodoro.tasks.data.local
 
 import com.taskodoro.storage.db.TaskodoroDB
-import com.taskodoro.storage.tasks.TaskStore
-import com.taskodoro.tasks.model.Task
+import com.taskodoro.tasks.data.TaskStore
+import com.taskodoro.tasks.feature.model.Task
 
 internal class SQLDelightTaskStore(
     database: TaskodoroDB,
