+++ taskodoro/src/commonMain/kotlin/com/taskodoro/tasks/feature/TaskRepository.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,10 +14,12 @@
  *    limitations under the License.
  */
 
-package com.taskodoro.storage.tasks
+package com.taskodoro.tasks.feature
 
-import com.taskodoro.tasks.model.Task
+import com.taskodoro.tasks.feature.model.Task
 
-interface TaskStore {
-    fun save(task: Task)
+interface TaskRepository {
+    object SaveFailed : Exception()
+
+    fun save(task: Task): Result<Unit>
 }
