+++ taskodoro/src/commonMain/kotlin/com/taskodoro/tasks/feature/model/Task.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package com.taskodoro.tasks.model
+package com.taskodoro.tasks.feature.model
 
 import com.taskodoro.model.Uuid
