+++ taskodoro/src/commonTest/kotlin/com/taskodoro/helpers/Tasks.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package com.taskodoro.helpers
 
-import com.taskodoro.tasks.model.Task
+import com.taskodoro.tasks.feature.model.Task
 
 internal fun anyTask(
     title: String = "A task title",
