+++ taskodoro/src/commonTest/kotlin/com/taskodoro/onboarding/data/OnboardingStoreTest.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package com.taskodoro.onboarding.data
 
-import com.taskodoro.onboarding.OnboardingStore
+import com.taskodoro.onboarding.feature.OnboardingStore
 import com.taskodoro.storage.preferences.test.TestPreferences
 import kotlin.test.AfterTest
 import kotlin.test.Test
