+++ taskodoro/src/nativeMain/kotlin/com/taskodoro/onboarding/OnboardingStoreFactory.native.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 package com.taskodoro.onboarding
 
 import com.taskodoro.onboarding.data.PreferencesOnboardingStore
+import com.taskodoro.onboarding.feature.OnboardingStore
 import com.taskodoro.storage.preferences.PreferencesFactory
 
 actual class OnboardingStoreFactory {
