+++ taskodoro/src/nativeMain/kotlin/com/taskodoro/tasks/data/TaskStoreFactory.native.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,10 +14,10 @@
  *    limitations under the License.
  */
 
-package com.taskodoro.storage.tasks
+package com.taskodoro.tasks.data
 
 import com.taskodoro.storage.db.DatabaseFactory
-import com.taskodoro.storage.tasks.store.SQLDelightTaskStore
+import com.taskodoro.tasks.data.local.SQLDelightTaskStore
 
 actual class TaskStoreFactory {
